+++ src/python/pants/backend/experimental/java/register.py
@@ -2,7 +2,6 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 from pants.backend.java import classpath, tailor
-from pants.backend.java import util_rules as java_util_rules
 from pants.backend.java.compile import javac
 from pants.backend.java.dependency_inference import (
     import_parser,
@@ -21,6 +20,7 @@ from pants.backend.java.target_types import (
 )
 from pants.backend.java.target_types import rules as target_types_rules
 from pants.backend.java.test import junit
+from pants.jvm import jdk_rules
 from pants.jvm import util_rules as jvm_util_rules
 from pants.jvm.goals import coursier
 from pants.jvm.resolve import coursier_fetch, coursier_setup
@@ -55,6 +55,6 @@ def rules():
         *dependency_inference_rules.rules(),
         *tailor.rules(),
         *jvm_util_rules.rules(),
-        *java_util_rules.rules(),
+        *jdk_rules.rules(),
         *target_types_rules(),
     ]
