+++ src/python/pants/backend/java/compile/javac.py
@@ -5,16 +5,13 @@ from __future__ import annotations
 
 import logging
 from dataclasses import dataclass
-from enum import Enum
 from itertools import chain
 
 from pants.backend.java.target_types import JavaSourceField
-from pants.backend.java.util_rules import JdkSetup
 from pants.core.goals.check import CheckRequest, CheckResult, CheckResults
 from pants.core.util_rules.archive import ZipBinary
 from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
 from pants.engine.addresses import Addresses
-from pants.engine.engine_aware import EngineAwareReturnType
 from pants.engine.fs import (
     EMPTY_DIGEST,
     AddPrefix,
@@ -28,6 +25,8 @@ from pants.engine.process import BashBinary, FallibleProcessResult, Process, Pro
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import CoarsenedTarget, CoarsenedTargets, FieldSet, Sources, Targets
 from pants.engine.unions import UnionRule
+from pants.jvm.compile import CompiledClassfiles, CompileResult, FallibleCompiledClassfiles
+from pants.jvm.jdk_rules import JdkSetup
 from pants.jvm.resolve.coursier_fetch import (
     CoursierLockfileForTargetRequest,
     CoursierResolvedLockfile,
@@ -36,7 +35,6 @@ from pants.jvm.resolve.coursier_fetch import (
 )
 from pants.jvm.resolve.coursier_setup import Coursier
 from pants.util.logging import LogLevel
-from pants.util.strutil import strip_v2_chroot_path
 
 logger = logging.getLogger(__name__)
 
@@ -55,78 +53,6 @@ class JavacCheckRequest(CheckRequest):
 @dataclass(frozen=True)
 class CompileJavaSourceRequest:
     component: CoarsenedTarget
-
-
-@dataclass(frozen=True)
-class CompiledClassfiles:
-    """The outputs of a compilation contained in either zero or one JAR file.
-
-    TODO: Rename this type to align with the guarantee about its content.
-    """
-
-    digest: Digest
-
-
-class CompileResult(Enum):
-    SUCCEEDED = "succeeded"
-    FAILED = "failed"
-    DEPENDENCY_FAILED = "dependency failed"
-
-
-@dataclass(frozen=True)
-class FallibleCompiledClassfiles(EngineAwareReturnType):
-    description: str
-    result: CompileResult
-    output: CompiledClassfiles | None
-    exit_code: int
-    stdout: str | None = None
-    stderr: str | None = None
-
-    @classmethod
-    def from_fallible_process_result(
-        cls,
-        description: str,
-        process_result: FallibleProcessResult,
-        output: CompiledClassfiles | None,
-        *,
-        strip_chroot_path: bool = False,
-    ) -> FallibleCompiledClassfiles:
-        def prep_output(s: bytes) -> str:
-            return strip_v2_chroot_path(s) if strip_chroot_path else s.decode()
-
-        exit_code = process_result.exit_code
-        # TODO: Coursier renders this line on macOS.
-        stderr = "\n".join(
-            line
-            for line in prep_output(process_result.stderr).splitlines()
-            if line != "setrlimit to increase file descriptor limit failed, errno 22"
-        )
-        return cls(
-            description=description,
-            result=(CompileResult.SUCCEEDED if exit_code == 0 else CompileResult.FAILED),
-            output=output,
-            exit_code=exit_code,
-            stdout=prep_output(process_result.stdout),
-            stderr=stderr,
-        )
-
-    def level(self) -> LogLevel:
-        return LogLevel.ERROR if self.exit_code != 0 else LogLevel.INFO
-
-    def message(self) -> str:
-        message = self.description
-        message += (
-            " succeeded." if self.exit_code == 0 else f" failed (exit code {self.exit_code})."
-        )
-        if self.stdout:
-            message += f"\n{self.stdout}"
-        if self.stderr:
-            message += f"\n{self.stderr}"
-        return message
-
-    def cacheable(self) -> bool:
-        # Failed compile outputs should be re-rendered in every run.
-        return self.exit_code == 0
 
 
 @rule(desc="Compile with javac")
